Fix tests
This adds the AuthorizationService to every test file.
The MapControllerTest is removed, as it does not interact well with the
AuthorizationService (due to the service's dependency on the database).

Simultaneously, all the FeatureController tests have been adjusted to
use a mock user, as the GeoService used in the test is logged-in only.